# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Slop Studios 3 API
  description: |
    AI-powered creative platform API for media generation, agent orchestration, and live coding music.

    ## Features

    - **Authentication**: JWT-based authentication with session management
    - **Claude Integration**: AI-powered text generation with streaming support
    - **Strudel Integration**: Live coding music with pattern validation and audio rendering
    - **Agent Orchestration**: Multi-agent workflows with various orchestration patterns
    - **Agent Collaboration**: Self-critique and multi-agent discussions
    - **Real-time Updates**: WebSocket support for streaming responses

    ## Authentication

    All authenticated endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    Tokens are obtained via the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.

    ## Rate Limiting

    API requests are rate-limited. When rate limited, responses include:
    - `429 Too Many Requests` status code
    - `Retry-After` header with seconds until retry is allowed

    ## Error Responses

    All error responses follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "category": "validation|authentication|authorization|not_found|conflict|rate_limit|client|external|internal|timeout",
        "details": {}
      },
...

...
servers:
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Creates a new user account and returns an authentication token.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticates a user and returns a JWT token.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Returns the profile of the currently authenticated user.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh token
      description: Refreshes the JWT token using the current valid token.
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================================================
  # Claude Integration Endpoints
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            category:
              type: string
              enum: [validation, authentication, authorization, not_found, conflict, rate_limit, client, external, internal, timeout]
            details:
              type: object
              additionalProperties: true
            retryAfter:
              type: integer
              description: Seconds until retry is allowed (for rate limits)
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # =========================================================================
    # Health Schemas
    # =========================================================================
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 100
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
            expiresIn:
              type: integer
    UserResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    TokenResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            expiresIn:
              type: integer
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # =========================================================================
    # Claude Schemas
    # =========================================================================
...